Hook:Implement custom hook for using locked route with react query and navigate
